requirements.txt
djangorestframework
psycopg2-binary
gunicorn
django-heroku